Comment and restructure code
